@@ -4,19 +4,19 @@
   "private": true,
   "proxy": "http://localhost:3000",
   "engines": {
-    "node": "16.x"
+    "node": "20.x"
   },
   "dependencies": {
-    "@testing-library/jest-dom": "^5.13.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "react": "^17.0.2",
-    "react-currency-input-field": "^3.6.10",
-    "react-dom": "^17.0.2",
+    "@testing-library/jest-dom": "^6.1.5",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.1",
+    "react": "^18.2.0",
+    "react-currency-input-field": "^3.10.0",
+    "react-dom": "^18.2.0",
     "react-intl-currency-input": "^0.2.6",
-    "react-router-dom": "^5.3.4",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.1.2"
+    "react-router-dom": "^6.20.1",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^3.5.0"
   },
   "scripts": {
     "start": "PORT=4000 react-scripts start",
