@@ -2,7 +2,7 @@
   "name": "phase-4-deploying-app-demo",
   "description": "Build scripts for Heroku",
   "engines": {
-    "node": "16.x"
+    "node": "20.x"
   },
   "scripts": {
     "build": "npm install --prefix client && npm run build --prefix client",
