@@ -7,19 +7,12 @@ class User < ApplicationRecord
   has_many :debts
   has_many :payment_debts, through: :debts, source: :payment
 
-  PASSWORD_REQUIREMENTS = /\A
-  (?=.{8,})          # Must contain 20 or more characters
-  (?=.*\d)            # Must contain a digit
-  (?=.*[a-z])         # Must contain a lowercase character
-  (?=.*[A-Z])         # Must contain an uppercase character
-  (?=.*[[:^alnum:]])  # Must contain a symbol
-/x
+  PASSWORD_REQUIREMENTS = /\A(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^a-zA-Z\d]).{8,}\z/
 
   #Validations
   validates :name, presence: true
   validates :email, presence: true, uniqueness: true, format: { with: 
   URI::MailTo::EMAIL_REGEXP }
   validates :password, presence: true, format: { with: PASSWORD_REQUIREMENTS, 
-  message: "must contain at least 8 characters, 1 number, 1 uppercase, 1 
-  speacial character" }
+  message: "must contain at least 8 characters, 1 uppercase, 1 lowercase, 1 number, and 1 special character" }
 end
